+++ src/main/java/com/powsybl/openloadflow/network/LfContingency.java
@@ -9,20 +9,15 @@ package com.powsybl.openloadflow.network;
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.powsybl.loadflow.LoadFlowParameters;
-import com.powsybl.openloadflow.graph.GraphDecrementalConnectivity;
 import com.powsybl.openloadflow.util.PerUnit;
-import com.powsybl.openloadflow.util.sa.PropagatedContingency;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.commons.lang3.tuple.Triple;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.io.Writer;
-import java.util.HashSet;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
@@ -43,7 +38,7 @@ public class LfContingency {
 
     private final Set<LfGenerator> generators;
 
-    private double activePowerLoss;
+    private double activePowerLoss = 0;
 
     public LfContingency(String id, int index, Set<LfBus> buses, Set<LfBranch> branches, Set<Pair<LfShunt, Double>> shunts,
                          Set<Triple<LfBus, Pair<Double, Double>, Double>> loadBuses, Set<LfGenerator> generators) {
@@ -54,17 +49,15 @@ public class LfContingency {
         this.shunts = Objects.requireNonNull(shunts);
         this.loadBuses = Objects.requireNonNull(loadBuses);
         this.generators = Objects.requireNonNull(generators);
-        double lose = 0;
         for (LfBus bus : buses) {
-            lose += bus.getGenerationTargetP() - bus.getLoadTargetP();
+            activePowerLoss += bus.getGenerationTargetP() - bus.getLoadTargetP();
         }
         for (Triple<LfBus, Pair<Double, Double>, Double> loadBus : loadBuses) {
-            lose -= loadBus.getMiddle().getKey();
+            activePowerLoss -= loadBus.getMiddle().getKey();
         }
         for (LfGenerator generator : generators) {
-            lose += generator.getTargetP();
+            activePowerLoss += generator.getTargetP();
         }
-        this.activePowerLoss = lose;
     }
 
     public String getId() {
@@ -85,67 +78,6 @@ public class LfContingency {
 
     public double getActivePowerLoss() {
         return activePowerLoss;
-    }
-
-    public static Optional<LfContingency> create(PropagatedContingency propagatedContingency, LfNetwork network,
-                                                 GraphDecrementalConnectivity<LfBus> connectivity, boolean useSmallComponents) {
-        // find contingency branches that are part of this network
-        Set<LfBranch> branches = new HashSet<>(1);
-        for (String branchId : propagatedContingency.getBranchIdsToOpen()) {
-            LfBranch branch = network.getBranchById(branchId);
-            if (branch != null) {
-                branches.add(branch);
-            }
-        }
-
-        // check if contingency split this network into multiple components
-        if (branches.isEmpty() && propagatedContingency.getShuntIdsToLose().isEmpty()
-                && propagatedContingency.getLoadIdsToLose().isEmpty() && propagatedContingency.getGeneratorIdsToLose().isEmpty()) {
-            return Optional.empty();
-        }
-
-        // update connectivity with triggered branches
-        for (LfBranch branch : branches) {
-            connectivity.cut(branch.getBus1(), branch.getBus2());
-        }
-
-        // add to contingency description buses and branches that won't be part of the main connected
-        // component in post contingency state
-        Set<LfBus> buses;
-        if (useSmallComponents) {
-            buses = connectivity.getSmallComponents().stream().flatMap(Set::stream).collect(Collectors.toSet());
-        } else {
-            int slackBusComponent = connectivity.getComponentNumber(network.getSlackBus());
-            buses = network.getBuses().stream().filter(b -> connectivity.getComponentNumber(b) != slackBusComponent).collect(Collectors.toSet());
-        }
-        buses.forEach(b -> branches.addAll(b.getBranches()));
-
-        Set<Pair<LfShunt, Double>> shunts = new HashSet<>(1);
-        for (Pair<String, Double> shuntAndB : propagatedContingency.getShuntIdsToLose()) {
-            LfShunt shunt = network.getShuntById(shuntAndB.getKey());
-            if (shunt != null) {
-                shunts.add(Pair.of(shunt, shuntAndB.getValue()));
-            }
-        }
-
-        Set<LfGenerator> generators = new HashSet<>(1);
-        for (Pair<String, Double> generatorInfo : propagatedContingency.getGeneratorIdsToLose()) {
-            LfGenerator generator = network.getGeneratorById(generatorInfo.getKey());
-            generators.add(generator);
-        }
-
-        Set<Triple<LfBus, Pair<Double, Double>, Double>> loadBuses = new HashSet<>(1);
-        for (Triple<String, Pair<Double, Double>, Double> loadInfo : propagatedContingency.getLoadIdsToLose()) {
-            LfBus bus = network.getBusById(loadInfo.getLeft());
-            if (bus != null) {
-                loadBuses.add(Triple.of(bus, loadInfo.getMiddle(), loadInfo.getRight()));
-            }
-        }
-
-        // reset connectivity to discard triggered branches
-        connectivity.reset();
-
-        return Optional.of(new LfContingency(propagatedContingency.getContingency().getId(), propagatedContingency.getIndex(), buses, branches, shunts, loadBuses, generators));
     }
 
     public void apply(LoadFlowParameters parameters) {
