+++ src/main/java/com/powsybl/openloadflow/sa/AcSecurityAnalysis.java
@@ -31,8 +31,8 @@ import com.powsybl.openloadflow.equations.EquationUtil;
 import com.powsybl.openloadflow.graph.GraphDecrementalConnectivity;
 import com.powsybl.openloadflow.network.*;
 import com.powsybl.openloadflow.network.impl.Networks;
+import com.powsybl.openloadflow.network.impl.PropagatedContingency;
 import com.powsybl.openloadflow.network.util.PreviousValueVoltageInitializer;
-import com.powsybl.openloadflow.util.sa.PropagatedContingency;
 import com.powsybl.security.*;
 import com.powsybl.security.monitor.StateMonitor;
 import com.powsybl.security.results.BranchResult;
@@ -140,7 +140,7 @@ public class AcSecurityAnalysis extends AbstractSecurityAnalysis {
                 GraphDecrementalConnectivity<LfBus> connectivity = network.createDecrementalConnectivity(connectivityProvider);
                 while (contingencyIt.hasNext()) {
                     PropagatedContingency propagatedContingency = contingencyIt.next();
-                    LfContingency.create(propagatedContingency, network, connectivity, true)
+                    propagatedContingency.toLfContingency(network, connectivity, true)
                             .ifPresent(lfContingency -> { // only process contingencies that impact the network
                                 lfContingency.apply(loadFlowParameters);
