+++ src/test/java/com/powsybl/openloadflow/util/sa/ContingencyTrippingTest.java
@@ -11,6 +11,7 @@ import com.powsybl.iidm.network.*;
 import com.powsybl.iidm.network.test.EurostagTutorialExample1Factory;
 import com.powsybl.iidm.network.test.FictitiousSwitchFactory;
 import com.powsybl.openloadflow.network.NodeBreakerNetworkFactory;
+import com.powsybl.openloadflow.network.impl.ContingencyTripping;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
